requirements.txt
numpy
pillow
scikit-learn
torch
torchvision
tqdm
argparse
opencv-python
Pillow 
pydicom
progress
datetime
tensorboard
pandas
customtkinter